Improve UART and radio code, add configuration display and user external callback for customize project.
Use RTX source code instead of library.
Add platform initialize return value for customer check error type.

+++ bsp/cpu/HDSC/HC32L13X/emdrv/bsp_lpower.h
@@ -43,6 +43,6 @@ void BSP_LPowerStop(void);
  *
  * @return  same as LPTimer IRQ count before stop.
  */
-uint16_t BSP_LPowerGetSecs(void);
+uint32_t BSP_LPowerGetSecs(void);
 
 #endif
